% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/signpostTestFunctions.R
\name{calc_p_val}
\alias{calc_p_val}
\title{Calculates p value for null distr and observed value}
\usage{
calc_p_val(thetaObs, gammaH0, thetasH0)
}
\arguments{
\item{thetaObs}{numeric in \link{0,1} test statistic}

\item{gammaH0}{numeric in \link{0,1} the true value of gamma under H0}

\item{thetasH0}{numeric in \link{0,1} the generated thetas under H0}
}
\value{
p value numeric in \link{0,1}
}
\description{
Calculates p value for null distr and observed value
}
