% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulationFunctions.R
\name{calculate_eta}
\alias{calculate_eta}
\title{Calculate eta}
\usage{
calculate_eta(X, theta, beta_0, beta_a)
}
\arguments{
\item{X}{design matrix of the penalized covariates.}

\item{theta}{numeric, in \link{0,1} weight parameter of targets beta_0 and beta_a}

\item{beta_0}{numeric, The number of elements should match the number of columns of X.}

\item{beta_a}{numeric, The number of elements should match the number of columns of X.}
}
\value{
eta, numeric with same number of elements as rows in X
}
\description{
this function calculates X * ((1-theta) beta_0 + theta . beta_a)
}
