% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/signpostTestFunctions.R
\name{create_progress_reporter}
\alias{create_progress_reporter}
\title{Create a progress reporter}
\usage{
create_progress_reporter(total, name = "Progress")
}
\description{
Create a progress reporter
}
\keyword{internal}
