% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulationFunctions.R
\name{generate_X}
\alias{generate_X}
\title{Sample/generate normal design matrix}
\usage{
generate_X(n, p, sd = 1)
}
\arguments{
\item{n}{positive integer, sample size}

\item{p}{positive integer, dimension}

\item{sd}{positive numeric, standard deviation of columns, DEFAULT = 1}
}
\value{
X design matrix of the penalized covariates.
}
\description{
Sample/generate normal design matrix
}
