% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulationFunctions.R
\name{generate_Y}
\alias{generate_Y}
\title{Generate Response Vector Y}
\usage{
generate_Y(eta, model, n, sigma = 1)
}
\arguments{
\item{eta}{A numeric vector representing the linear predictor.}

\item{model}{A character string specifying the model type. Must be one of
\code{"linear"}, \code{"logistic"}, or \code{"poisson"}.}

\item{n}{A positive integer indicating the sample size. Must be equal to \code{length(eta)}.}

\item{sigma}{Optional. A positive numeric value specifying the standard deviation of the
noise for the linear model. Default is 1. Only used when \code{model = "linear"}.}
}
\value{
A numeric vector containing the generated responses.
}
\description{
This function generates a response vector Y from a given linear predictor (eta)
according to the specified model. For a logistic model, it generates binary responses
using a logistic link; for a Poisson model, it generates count data; and for a linear model,
it adds normally distributed error to eta.
}
\examples{
# Logistic model example:
eta <- rnorm(100)
Y_logistic <- generate_Y(eta, model = "logistic", n = 100)

# Poisson model example:
eta <- rnorm(100)
Y_poisson <- generate_Y(eta, model = "poisson", n = 100)

# Linear model example:
eta <- rnorm(100)
Y_linear <- generate_Y(eta, model = "linear", n = 100, sigma = 2)

}
