% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/database_functions.R
\name{get_or_insert_param_id}
\alias{get_or_insert_param_id}
\title{Get or Insert Parameter ID}
\usage{
get_or_insert_param_id(
  con,
  n,
  p,
  lambda,
  test_type,
  GLM_model,
  B_null_distr,
  X,
  beta,
  n_beta = NULL,
  lambda_beta = NULL
)
}
\arguments{
\item{con}{A database connection object.}

\item{n}{Integer, the number of samples.}

\item{p}{Integer, the number of predictors.}

\item{lambda}{Numeric, regularization parameter.}

\item{test_type}{Character, type of test used.}

\item{GLM_model}{Character, the type of GLM model.}

\item{B_null_distr}{Integer, number of bootstrap samples.}

\item{X}{Character, representation of predictor variables.}

\item{beta}{Character, representation of model coefficients.}

\item{n_beta}{(Optional) Integer, sample size used to estimate beta_0 and beta_a.}

\item{lambda_beta}{(Optional) Numeric, ridge penalty used when estimating beta_0 and beta_a.}
}
\value{
Integer, the \code{id} corresponding to the given parameter set.
}
\description{
Retrieves the ID of an existing parameter set in the \code{parameters} table.
If the parameters do not exist, inserts them and returns the new ID.
}
