% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/database_functions.R
\name{get_results_by_params}
\alias{get_results_by_params}
\title{Retrieve Simulation Results for Given Parameters}
\usage{
get_results_by_params(
  con,
  n,
  p,
  lambda,
  test_type,
  GLM_model,
  B_null_distr,
  X,
  beta
)
}
\arguments{
\item{con}{A database connection object.}

\item{n}{Integer, the number of samples.}

\item{p}{Integer, the number of predictors.}

\item{lambda}{Numeric, regularization parameter.}

\item{test_type}{Character, type of test used.}

\item{GLM_model}{Character, the type of GLM model.}

\item{B_null_distr}{Integer, number of bootstrap samples.}

\item{X}{Character, representation of predictor variables.}

\item{beta}{Character, representation of model coefficients.}
}
\value{
A data frame containing the simulation results for the given parameters.
}
\description{
Queries the \code{results} table to retrieve results corresponding to a given parameter set.
}
