% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/signpostTestFunctions.R
\name{penalty_theta}
\alias{penalty_theta}
\title{Calculates penalty}
\usage{
penalty_theta(theta, Y, X, beta_0, beta_a, lambda, model = "logistic")
}
\arguments{
\item{theta}{numeric, in \link{0,1} weight parameter of targets beta_0 and beta_a}

\item{Y}{numeric, response vector}

\item{X}{design matrix of the penalized covariates, The number of rows should
match the number of elements of Y.}

\item{beta_0}{numeric, towards which the estimate is shrunken. The number of
elements should match the number of columns of X.}

\item{beta_a}{numeric, towards which the estimate is shrunken. The number of
elements should match the number of columns of X.}

\item{lambda}{positive numeric, the regular ridge penalty parameter}

\item{model}{string specifying "linear", "logistic" or "poisson"}
}
\value{
penalty numeric
}
\description{
Calculates penalty
}
