% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulationFunctions.R
\name{power_LR_test_extended}
\alias{power_LR_test_extended}
\title{Extended power analysis for likelihood ratio test}
\usage{
power_LR_test_extended(
  n,
  p,
  model,
  lambda,
  B_power,
  n_beta = NULL,
  lambda_beta = NULL,
  estimate_beta = FALSE,
  lambda_est = lambda_est,
  misspecified = FALSE,
  scaled = TRUE
)
}
\arguments{
\item{n}{Positive integer, sample size}

\item{p}{Positive integer, dimension}

\item{model}{Character specifying "linear", "logistic", or "poisson"}

\item{lambda}{Positive numeric, ridge penalty parameter}

\item{B_power}{Positive integer, number of simulation iterations}

\item{n_beta}{Positive integer, sample size for beta estimation}

\item{lambda_beta}{Positive numeric, penalty for beta estimation}

\item{estimate_beta}{Logical, whether to estimate beta values}

\item{lambda_est}{Numeric, regularization parameter for loss estimation}

\item{misspecified}{Logical, whether to use misspecified beta_0}

\item{scaled}{Logical, whether to scale estimated beta_a}
}
\value{
List containing p_values matrix
}
\description{
Extended power analysis for likelihood ratio test
}
