% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulationFunctions.R
\name{power_signpost_test}
\alias{power_signpost_test}
\title{Calculates power of signpost test - RESAMPLE}
\usage{
power_signpost_test(
  n,
  p,
  gammaH0,
  model,
  lambda,
  B_X,
  B_theta_null,
  B_power,
  n_cores = 8,
  n_beta = NULL,
  lambda_beta = NULL,
  estimate_beta = FALSE
)
}
\arguments{
\item{n}{Positive integer, sample size.}

\item{p}{Positive integer, dimension (length of beta_0 and beta_a).}

\item{gammaH0}{Numeric in \link{0,1}, specifying the null hypothesis value.}

\item{model}{Character, specifying "linear", "logistic", or "poisson".}

\item{lambda}{Positive numeric, regular ridge penalty parameter.}

\item{B_X}{Numeric, number of X's to be generated.}

\item{B_theta_null}{Numeric, size of the null distribution (per test).}

\item{B_power}{Numeric, number of p-values per gamma per generated X.}

\item{n_cores}{Optional integer, number of cores for parallel programming.}

\item{n_beta}{Positive integer, sample size to estimate beta_0 and beta_a.}

\item{lambda_beta}{Positive numeric, regular ridge penalty parameter when estimating beta_0 and beta_a.}

\item{estimate_beta}{Logical, if TRUE  beta_a are estimated; if FALSE, true values are used.}
}
\value{
A matrix of p-values with \code{B_X * B_power} rows and \code{length(gammas)} columns.
}
\description{
For gamma = 0, 0.1, ..., 0.5
Signpost test: H_0: gamma = gammaH0 vs gamma != gammaH0
}
