% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulationFunctions.R
\name{power_signpost_test_extended_AS}
\alias{power_signpost_test_extended_AS}
\title{Extended power analysis for signpost test with asymptotic standard errors}
\usage{
power_signpost_test_extended_AS(
  n,
  p,
  gammaH0,
  model,
  lambda,
  B_X,
  B_power,
  n_cores = 8,
  n_beta = NULL,
  lambda_beta = NULL,
  estimate_beta = FALSE,
  use_theta_hat_in_variance = FALSE,
  fisher = FALSE,
  hybrid = FALSE,
  term1ONLY = FALSE,
  lambda_est = lambda_est,
  misspecified = FALSE,
  scaled = TRUE
)
}
\arguments{
\item{n}{Positive integer, sample size.}

\item{p}{Positive integer, dimension.}

\item{gammaH0}{Numeric in \link{0,1}, specifying the null hypothesis value.}

\item{model}{Character, specifying "linear", "logistic", or "poisson".}

\item{lambda}{Positive numeric, regular ridge penalty parameter.}

\item{B_X}{Numeric, number of X's to be generated.}

\item{B_power}{Numeric, number of p-values per gamma per generated X.}

\item{n_cores}{Optional integer, number of cores for parallel programming.}

\item{n_beta}{Positive integer, sample size to estimate beta_0 and beta_a.}

\item{lambda_beta}{Positive numeric, regular ridge penalty parameter when estimating beta_0 and beta_a.}

\item{estimate_beta}{Logical, if TRUE beta_0 and beta_a are estimated; if FALSE, true values are used.}

\item{use_theta_hat_in_variance}{Logical, whether to use theta_hat in variance calculation.}

\item{fisher}{Logical, whether to use Fisher information for standard error.}

\item{hybrid}{Logical, reserved for future use}

\item{term1ONLY}{Logical, whether to use only first term in variance calculation}

\item{lambda_est}{Numeric, regularization parameter for loss estimation}

\item{misspecified}{Logical, whether to use misspecified beta_0 (all zeros)}

\item{scaled}{Logical, whether to scale estimated beta_a}
}
\value{
List containing:
\item{p_values}{Matrix of p-values}
\item{theta_hats}{Matrix of theta estimates}
\item{theta_hat_loss}{Matrix of losses using theta_hat}
\item{theta_null_loss}{Matrix of losses using theta=0}
}
\description{
Extended power analysis for signpost test with asymptotic standard errors
}
