% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulationFunctions.R
\name{power_signpost_test_extended_resample}
\alias{power_signpost_test_extended_resample}
\title{Extended power analysis for signpost test with resampling}
\usage{
power_signpost_test_extended_resample(
  n,
  p,
  gammaH0,
  model,
  lambda,
  B_X,
  B_theta_null,
  B_power,
  n_cores = 8,
  n_beta = NULL,
  lambda_beta = NULL,
  estimate_beta = FALSE,
  lambda_est = lambda_est
)
}
\arguments{
\item{n}{Positive integer, sample size.}

\item{p}{Positive integer, dimension (length of beta_0 and beta_a).}

\item{gammaH0}{Numeric in \link{0,1}, specifying the null hypothesis value.}

\item{model}{Character, specifying "linear", "logistic", or "poisson".}

\item{lambda}{Positive numeric, regular ridge penalty parameter.}

\item{B_X}{Numeric, number of X's to be generated.}

\item{B_theta_null}{Numeric, size of the null distribution (per test).}

\item{B_power}{Numeric, number of p-values per gamma per generated X.}

\item{n_cores}{Optional integer, number of cores for parallel programming.}

\item{n_beta}{Positive integer, sample size to estimate beta_0 and beta_a.}

\item{lambda_beta}{Positive numeric, regular ridge penalty parameter when estimating beta_0 and beta_a.}

\item{estimate_beta}{Logical, if TRUE  beta_a are estimated; if FALSE, true values are used.}

\item{lambda_est}{Numeric, regularization parameter for loss estimation}
}
\value{
List containing:
\item{p_values}{Matrix of p-values}
\item{theta_hats}{Matrix of theta estimates}
\item{theta_hat_loss}{Matrix of losses using theta_hat}
\item{theta_null_loss}{Matrix of losses using theta=0}
}
\description{
Performs power analysis and computes additional metrics including theta estimates
and loss functions for both theta_hat and theta=0.
}
