% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/signpostTestFunctions.R
\name{pval2power}
\alias{pval2power}
\title{Maps array of p_values to the power}
\usage{
pval2power(pval, alpha = 0.05)
}
\arguments{
\item{pval}{Array of p_values (numeric)}

\item{alpha}{numeric, default = 0.05}
}
\value{
pval
}
\description{
Maps array of p_values to the power
}
