% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/signpostTestFunctions.R
\name{signpostTestGLM}
\alias{signpostTestGLM}
\title{Signpost test - RESAMPLE}
\usage{
signpostTestGLM(
  Y,
  X,
  beta_0,
  beta_a,
  lambda,
  gammaH0 = 0,
  B = 1000,
  model = "logistic"
)
}
\arguments{
\item{Y}{numeric, response vector}

\item{X}{design matrix of the penalized covariates, The number of rows should
match the number of elements of Y.}

\item{beta_0}{numeric, towards which the estimate is shrunken. The number of
elements should match the number of columns of X.}

\item{beta_a}{numeric, towards which the estimate is shrunken. The number of
elements should match the number of columns of X.}

\item{lambda}{positive numeric, the regular ridge penalty parameter}

\item{gammaH0}{numeric in \link{0,1} the true value of gamma under H0}

\item{B}{numeric, specifying the to be generated size of the
null distribution (per test)}

\item{model}{string specifying "linear", "logistic" or "poisson"}
}
\value{
pvalue numeric
}
\description{
Signpost test - RESAMPLE
}
