% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/signpostTestFunctions.R
\name{theta_hat_lambda}
\alias{theta_hat_lambda}
\title{Calculates theta_hat for given lambda < Inf with one-time random rotation and optimization}
\usage{
theta_hat_lambda(
  Y,
  X,
  beta_0,
  beta_a,
  lambda,
  model = c("logistic", "poisson", "linear"),
  do_rotation = FALSE,
  grid_length = NULL
)
}
\arguments{
\item{Y}{Numeric, response vector.}

\item{X}{Design matrix of the penalized covariates; the number of rows should match the length of Y.}

\item{beta_0}{Numeric, shrinkage target; length must match the number of columns of X.}

\item{beta_a}{Numeric, alternative target; length must match the number of columns of X.}

\item{lambda}{Positive numeric, the regular ridge penalty parameter.}

\item{model}{Character, one of "linear", "logistic", or "poisson". Defaults to "logistic".}

\item{do_rotation}{Logical, default FALSE. If TRUE, applies a random orthonormal rotation to (beta_a - beta_0) once.}

\item{grid_length}{Optional integer; if provided, a grid search with that many points will be performed instead of using optimize().}
}
\value{
theta_hat Numeric in \link{0,1}, the estimated optimal theta.
}
\description{
Calculates theta_hat for given lambda < Inf with one-time random rotation and optimization
}
