% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/signpostTestFunctions.R
\name{theta_inf_hat}
\alias{theta_inf_hat}
\title{Computes theta_hat for lambda = infinity}
\usage{
theta_inf_hat(Y, X, U, beta_0, beta_a, model, do_rotation = FALSE)
}
\arguments{
\item{Y}{Numeric vector, response variable.}

\item{X}{Numeric matrix, design matrix of penalized covariates. Rows should match the length of Y.}

\item{U}{Numeric matrix, design matrix of unpenalized covariates. Rows should match the length of Y.}

\item{beta_0}{Numeric vector, shrinkage target (same length as columns of X).}

\item{beta_a}{Numeric vector, alternative hypothesis shrinkage target (same length as columns of X).}

\item{model}{String, one of "linear", "logistic", or "poisson".}

\item{do_rotation}{Logical, whether to apply random rotation.}
}
\value{
Numeric in \link{0,1}, the estimated theta_hat.
}
\description{
This function calculates the estimated theta_hat for lambda set to infinity.
}
