% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/signpostTestFunctions.R
\name{theta_Y}
\alias{theta_Y}
\title{Generates Y and calculates theta_hat for given lambda (possibly inf)}
\usage{
theta_Y(X, beta_0, beta_a, eta, model, lambda)
}
\arguments{
\item{X}{design matrix of the penalized covariates, The number of rows should
match the number of elements of Y.}

\item{beta_0}{numeric, towards which the estimate is shrunken. The number of
elements should match the number of columns of X.}

\item{beta_a}{numeric, towards which the estimate is shrunken. The number of
elements should match the number of columns of X.}

\item{eta}{numeric linear predictor with same number of elements as rows in X}

\item{model}{string specifying "linear", "logistic" or "poisson"}

\item{lambda}{positive numeric, the regular ridge penalty parameter}
}
\value{
theta_hat numeric in \link{0,1}
}
\description{
Generates Y and calculates theta_hat for given lambda (possibly inf)
}
