% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/signpostTestFunctions.R
\name{thetaInf}
\alias{thetaInf}
\title{Helper function to theta_inf_hat, calculates RHS of estimating equation}
\usage{
thetaInf(theta, Y, X, U, lpTarget0, lpTargetD, model)
}
\arguments{
\item{theta}{Numeric in \link{0,1}, shrinkage parameter.}

\item{Y}{Numeric vector, response variable.}

\item{X}{Numeric matrix, design matrix of penalized covariates. Rows should match length of Y.}

\item{U}{Numeric matrix, design matrix of unpenalized covariates. Rows should match length of Y.}

\item{lpTarget0}{Precomputed numeric vector, linear predictor for beta_0.}

\item{lpTargetD}{Precomputed numeric vector, linear predictor difference for beta_a - beta_0.}

\item{model}{String, one of "linear", "logistic", or "poisson".}
}
\value{
Numeric, right-hand side of estimating equation.
}
\description{
Helper function to theta_inf_hat, calculates RHS of estimating equation
}
