% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/signpostTestFunctions.R
\name{validate_signpost_inputs}
\alias{validate_signpost_inputs}
\title{Validate common inputs for signpost test functions}
\usage{
validate_signpost_inputs(Y, X, beta_0, beta_a, model)
}
\description{
Validate common inputs for signpost test functions
}
\keyword{internal}
